1205 : changes based on comments
